requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
pydantic>=2.5
pyyaml>=6.0
typer>=0.9
rich>=13.0
joblib>=1.3
streamlit>=1.33
